+++ catalog-service/src/main/resources/data.sql
@@ -6,7 +6,7 @@ CREATE TABLE product (
   title VARCHAR(250) NOT NULL,
   description VARCHAR(250) NOT NULL,
   brand VARCHAR(25) NOT NULL,
-  price INT NOT NULL,
+  price DECIMAL(20, 2) NOT NULL,
   color VARCHAR(25) NOT NULL
 );
